requirements.txt
aiohttp==3.7.4.post0
anyio==3.4.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.3
discord.py @ git+https://github.com/Rapptz/discord.py@45d498c1b76deaf3b394d17ccf56112fa691d160
h11==0.12.0
idna==3.3
multidict==5.2.0
python-dotenv==0.19.2
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.0
uvicorn==0.15.0
yarl==1.7.2
